Update patch set 1
Patch Set 1: Do not submit

Yeah :) Skyrocket and Hercules are OK with GLES because of their low resolution. Move up to Quincy and it's a jankfest. What's happening is that it's unable to use hardware layer composition (MDP) and falling back to OpenGL, which is really bad. The new 4-layer composition stuff that was done for 8960 and above that's in the display-caf (it's the reason why it exists, actually) doesn't work properly on any 8660s that we've tested so far. 

We need to figure out why it's failing in the kernel. It seems that get_img is expecting to find a shared ION FD, but it's not there.

Patch-set: 1