Update patch set 1
Patch Set 1: Code-Review-1

Have you measured it with a power meter?

It's likely that you'll actually see increased drain, due to the cost of ramping up from a lower frequency.

The behavior where the CPU hangs out at a higher frequency is by design and is controlled by our power HAL which handles scroll and touch events. The CPU freq is boosted to max for a specific amount of time since we know the device is actively being used and requires a high level of interactivity. It also avoids constant ramp up/ramp down which is costly.

Patch-set: 1
Label: Code-Review=-1