Update patch set 1
Patch Set 1:

This affects a number of devices, and what actually happens is that "performance" is the default governor when the system is powered up in order to boot as quickly as possible. After it's booted, we switch to interactive.

But the way this works is a total clusterfuck and it conflicts with the built-in CPU settings. We need to fix that.

Patch-set: 1