Update patch set 1
Patch Set 1: Code-Review-2

Init should be getting a uevent from the kernel when the governor changes, and refreshing the permissions. If not, then we have a bug.

I'm OK with tuning the screen off code though. I wonder if adding a hint that says music is playing would be smarter so that we can save power when it's not.

Patch-set: 1
Label: Code-Review=-2