Update patch set 5
Patch Set 5: Verified+1 Code-Review+1

Updated the commit msg with an explanation of my findings and reasoning for this patch.

If you want me to move this out of qcom-common to msm8660-common, no problem. But we will have to do the same patch to msm8960-common for d2 and jf.

Let me know if you have any questions.

Patch-set: 5