Update patch set 5
Patch Set 5:

The suggested patch to healthd (ignoring values) is wrong. Like David explained above, the gauge needs to be narrowed down to the correct source, not filtered for bad ones.

And... even if it was a matter of filtering out the wrong inputs instead of selecting the correct one, the libhealth override would still be the proper way to do it.

Patch-set: 5