Update patch set 2
Patch Set 2:

Of course it will only apply to newly built recoveries, *if* the recovery patch is merged. :)

As it stands though, this is a no-op without the relevant patch to recovery.

So either merge the recovery patch and don't merge this, or merge this and not the recovery patch. Simple. :)

Both patches are already in CM11 and working fine there, but there's one huge problem with CM11 built recoveries: You can't use it to install previous CM versions. Hence I'm not even slightly considering building it for my users.

Patch-set: 2
Reviewer: Arne Coucheron <444@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0