Update patch set 1
Patch Set 1:

Steve,

You added media_codecs_8974.xml to this repo but it's obviously not being used (if it was, the build would fail due to this typo).

It doesn't seem to be getting used for me and my device repo (mondrianwifi) because TARGET_BOARD_PLATFORM is not defined when the qcom-common.mk is included.

Two questions:

1. Should we be using the alternative media_codecs for msm8974 devices?  If so, the dependent commit should be merged as well (adding ffmpeg codecs).

2. How should I fix the lack of TARGET_BOARD_PLATFORM?  I hacked it to make sure things worked as: http://review.cyanogenmod.org/#/c/63524/ but that really doesn't seem right.

Patch-set: 1