Update patch set 1
Patch Set 1: Code-Review+1

I built jflte to confirm that the script is still included in its build.

FYI, before we moved to msm8974-common, the msm8974 devices were installing this script and thermald-8974.conf.  The script then links to that config file.  After moving to msm8974-common, we are no longer including ANY config for thermald.  Aka, thermald fails to start at all.

To fix this for mondrianwifi, I opted to include my own thermald.conf.  This seems reasonable (to me) to keep as a per device configuration at this point.  See:

http://review.cyanogenmod.org/#/c/68689/

You'll probably want to either disagree and add it to msm8974 common or agree and add the appropriate config to hlte.

Patch-set: 1
Label: Code-Review=+1