Update patch set 4
Patch Set 4:

Phyto: Yes, if you put the script into the main ramdisk, it will also get picked up in the recovery ramdisk.  If everyone agrees they want to do this, I'm fine with it.  It certainly won't harm anything to have the script in the main ramdisk and it will alleviate the need for the patch that I posted yesterday.

Long term (probably CM 12.0) I would like to get rid of the postrecoveryboot.sh scripts entirely and switch to a custom BCB implementation that recovery can manage properly.  For example, see http://review.cyanogenmod.org/68219 and http://review.cyanogenmod.org/68221 .

Patch-set: 4