Update patch set 3
Patch Set 3:

Ambient display is by default enabled, so on a fresh install, there will be a mismatch between the state in display settings and here. Enabled in display and disabled here. So probably best to keep it enabled here as well.

Other than that, I'm not too keen on the idea of having the option in two different places. Kind of redundant really. Would be much nicer and cleaner if it all could be consolidated into one ambient display category, but I don't know how much we can do as it is. Maybe try getting the Cyngn guys involved in doing something with it? :)

Patch-set: 3