Update patch set 1
Patch Set 1:

Hi Arne, I just tested it and yes it does charge with or without the new 'on charger' and 'service charger' stuff I put in, so I am not sure if the 'on charger' stuff is even needed, the only problem is that it's still blank, do you have any links to the display related stuff that I could refer to? Thanks.

Patch-set: 1