Update patch set 1
Patch Set 1: -Code-Review

The pickup part works two way : Pickup from a table, face-down, will use proximity sensor (reliable). The pickup face-up will rely on the accelerometer, and the issue is not about sensitivity but about sleep. If you check on the internet, a good portion of the Android devices have an Accelerometer that goes in deep-sleep with the CPU, unlike the proximity sensor. That issue causes the accelerometer not to deliver data. If you are charging the phone or connected to a PC, the pickup works just fine. I originally did it of the Xperia SP (huashan, + we have a lot more sensors than usual and none give a better detection), and the result is consistent since more than a week : all features working great expect face-up pickup when the phone is in deep-sleep, which actually may be a good thing after all.

Patch-set: 1
Reviewer: Adrian DC <13589@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=0